+++ src/vue/modules/requests/create-sale-requests/index.vue
@@ -8,8 +8,8 @@
           </template>
           <create-sale-form-module
             :request-id="selectedRequest.id"
-            @close="isDrawerShown = false"
-            @request-updated="initFirstPageLoader"
+            @submitted="isDrawerShown = false"
+            @request-updated="emitUpdateList('createSaleRequests:updateList')"
           />
         </template>
 
@@ -20,7 +20,7 @@
           <request-viewer
             :request="selectedRequest"
             @update-click="isUpdateMode = true"
-            @cancel="(isDrawerShown = false) || initFirstPageLoader()"
+            @cancel="closeDrawerAndUpdateList()"
           />
         </template>
       </drawer>
@@ -59,6 +59,7 @@ import { mapGetters, mapMutations, mapActions } from 'vuex'
 import { types } from './store/types'
 
 import { ErrorHandler } from '@/js/helpers/error-handler'
+import UpdateList from '@/vue/mixins/update-list.mixin'
 
 export default {
   name: 'create-sale-requests-module',
@@ -69,6 +70,8 @@ export default {
     RequestViewer,
     CreateSaleFormModule,
   },
+
+  mixins: [UpdateList],
 
   data: _ => ({
     isLoaded: false,
@@ -87,6 +90,11 @@ export default {
 
   async created () {
     this.initFirstPageLoader()
+    this.listenUpdateList('createSaleRequests:updateList', this.initFirstPageLoader)
+  },
+
+  beforeDestroy () {
+    this.resetUpdateListEvent('createSaleRequests:updateList')
   },
 
   methods: {
@@ -119,6 +127,11 @@ export default {
       this.isUpdateMode = false
       this.selectedRequest = request
       this.isDrawerShown = true
+    },
+
+    closeDrawerAndUpdateList () {
+      this.isDrawerShown = false
+      this.emitUpdateList('createSaleRequests:updateList')
     },
   },
 }