+++ src/vue/pages/Sales.vue
@@ -37,6 +37,21 @@
         </button>
       </template>
 
+      <!-- eslint-disable max-len -->
+      <template
+        v-else-if="getModule().canRenderSubmodule(CreateSaleFormModuleSimplified)"
+        slot="extra"
+      >
+        <button
+          v-ripple
+          class="app__button-raised"
+          @click="isCreateSaleDrawerShown = true"
+        >
+          <i class="mdi mdi-plus sales__btn-icon" />
+          {{ 'sales.create-sale' | globalize }}
+        </button>
+      </template>
+
       <template
         v-if="getModule().canRenderSubmodule(CreateOpportunityModule)"
         slot="extra"
@@ -62,7 +77,7 @@
         </template>
         <submodule-importer
           :submodule="getModule().getSubmodule(CreateOpportunityModule)"
-          @close="isAssetSaleDrawerShown = false"
+          @submitted="closeAssetSaleDrawerAndUpdateList()"
           :account-id="accountId"
           :min-amount="MIN_AMOUNT"
           :max-amount="MAX_AMOUNT"
@@ -82,6 +97,23 @@
         </template>
         <submodule-importer
           :submodule="getModule().getSubmodule(CreateSaleFormModule)"
+          @submitted="closeCreateSaleDrawerAndUpdateList()"
+        />
+      </drawer>
+    </template>
+
+    <!-- eslint-disable-next-line max-len -->
+    <template v-else-if="getModule().canRenderSubmodule(CreateSaleFormModuleSimplified)">
+      <drawer
+        :is-shown.sync="isCreateSaleDrawerShown"
+        :close-by-click-outside="false"
+        class="sales__drawer"
+      >
+        <template slot="heading">
+          {{ 'sales.new-sale' | globalize }}
+        </template>
+        <submodule-importer
+          :submodule="getModule().getSubmodule(CreateSaleFormModuleSimplified)"
           @close="isCreateSaleDrawerShown = false"
         />
       </drawer>
@@ -101,10 +133,12 @@ import { vuexTypes } from '@/vuex'
 import { vueRoutes } from '@/vue-router/routes'
 
 import { CreateSaleFormModule } from '@modules/create-sale-form/module'
+import { CreateSaleFormModuleSimplified } from '@modules/create-sale-form-simplified/module'
 
 import SubmoduleImporter from '@/modules-arch/submodule-importer'
 import { CreateOpportunityModule } from '@/vue/modules/create-opportunity/module'
 import { SalesListOwnedPageModule } from '@/vue/pages/sales/user-owned-sales-page-module'
+import UpdateList from '@/vue/mixins/update-list.mixin'
 
 export default {
   name: 'sales',
@@ -114,6 +148,8 @@ export default {
     SubmoduleImporter,
   },
 
+  mixins: [UpdateList],
+
   data: _ => ({
     isCreateSaleDrawerShown: false,
     isAssetSaleDrawerShown: false,
@@ -122,6 +158,7 @@ export default {
     MAX_AMOUNT: config.MAX_AMOUNT,
     DECIMAL_POINTS: config.DECIMAL_POINTS,
     CreateSaleFormModule,
+    CreateSaleFormModuleSimplified,
     vueRoutes,
     CreateOpportunityModule,
     SalesListOwnedPageModule,
@@ -131,6 +168,18 @@ export default {
     ...mapGetters({
       accountId: vuexTypes.accountId,
     }),
+  },
+
+  methods: {
+    closeAssetSaleDrawerAndUpdateList () {
+      this.isAssetSaleDrawerShown = false
+      this.emitUpdateList('sales:updateList')
+    },
+
+    closeCreateSaleDrawerAndUpdateList () {
+      this.isCreateSaleDrawerShown = false
+      this.emitUpdateList('sales:updateList')
+    },
   },
 }
 </script>