+++ CHANGELOG.md
@@ -7,15 +7,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
 
-## Unreleased
+## [Unreleased]
+#### Added
+- Real time updates on all pages
+
+#### Fixed
+- Bug about maximum call stack in buildKycRecoveryPageGuard
 
 #### Removed
 - Recovery seed screen on sign up 
 - Remove recovery signer from KYC recovery request building
 - Remove recovery seed screens on auth
 
-#### Fixed
-- Bug about maximum call stack in buildKycRecoveryPageGuard
+### Experimental features changes
+#### Added
+- Simplify form of create new sale (Conto)
+- Simplify pages for sale details (Conto)
 
 ## [1.10.0-rc.0] - 2019-07-16
 #### Added